BUG: FindFeatureReferenceCAxisOrientation - Use doubles to accumulate the StdDev values
Output values now agree with DREAM.3D > 6.5.172

+++ src/Plugins/OrientationAnalysis/src/OrientationAnalysis/Filters/Algorithms/ComputeFeatureReferenceCAxisMisorientations.cpp
@@ -167,19 +167,18 @@ Result<> ComputeFeatureReferenceCAxisMisorientations::operator()()
 
   // These 2 loops compute the population standard deviation of those misorientations for
   // each feature.
-  std::vector<float32> stdevs(totalFeatures, 0.0f);
-
+  std::vector<double> stdevs(totalFeatures, 0.0);
   for(usize cellIdx = 0; cellIdx < totalPoints; cellIdx++)
   {
     const int32 featureId = featureIds[cellIdx];
-    float diff = cellRefCAxisMis.getValue(cellIdx) - featAvgCAxisMis.getValue(featureId);
+    double diff = cellRefCAxisMis.getValue(cellIdx) - featAvgCAxisMis.getValue(featureId);
     stdevs[featureId] += (diff * diff);
   }
 
   // Finish computing the standard deviation in this loop
   for(usize featureId = 1; featureId < totalFeatures; featureId++)
   {
-    featStdevCAxisMis[featureId] = std::sqrt(stdevs[featureId] / static_cast<float32>(counts[featureId]));
+    featStdevCAxisMis[featureId] = std::sqrt(stdevs[featureId] / static_cast<double>(counts[featureId]));
   }
 
   return {};
